Update DZ 04.06.2025
